pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "evovir"
version = "0.1.0"
description = "Vertebrate host-range classifier built on Evo 2 DNA embeddings"
requires-python = ">=3.10"
dependencies = [
    "evo2",
    "torch>=2.1",
    "biopython>=1.83",
    "scikit-learn>=1.4",
    "pandas>=2.0",
    "numpy>=1.26",
    "tqdm",
    "pyyaml",
    "h5py",
    "requests",
    "matplotlib",
    "seaborn",
]

[project.optional-dependencies]
dev = ["pytest", "ipykernel", "jupyterlab"]

[tool.setuptools.packages.find]
where = ["."]
include = ["evovir*"]
